requirements.txt
openai
flask
pdfplumber
python-docx
spacy
python-dotenv
